error middleware, Logging middleware, authentication and login API

+++ api/services/user/user.controller.js
@@ -1,7 +1,31 @@
 const { Op } = require("sequelize");
 const { sequelize } = require("sequelize");
 const databaseInstance = require("../../utils/db/models/db_instance");
-// const handleApiError = require("../../../utils/middlewares/ErrorHandler");
+const handleApiError = require("../../utils/middlewares/ErrorHandler");
+const jwt = require("jsonwebtoken");
+require('dotenv').config();
+
+
+
+exports.Login = async (req, res) => {
+  try {
+    // Get admin user by request email
+    const user = await databaseInstance.db.User.findOne({ where: { Email: req.body.Email } });
+    if (!user) {
+      return res.status(401).send({ message: "Email is incorrect" });
+    }
+
+    // Retrieve token
+    const token = jwt.sign({ user: user.toJSON() }, process.env.HASHKEY, { expiresIn: '7d' });
+
+    res.status(200).send({ userInfo: user, token: token });
+  } catch (error) {
+    console.error(error);
+    handleApiError(res, error, "Login");
+  }
+};
+
+
 
 
 exports.getAllUsers = async (req, res) => {
@@ -11,6 +35,8 @@ exports.getAllUsers = async (req, res) => {
     res.status(200).send(users);
   } catch (error) {
     console.log(error);
+    handleApiError(res, error, "getAllUsers");
+
   }
 };
 
@@ -18,6 +44,8 @@ exports.getAllUsers = async (req, res) => {
 exports.getUserById = async (req, res) => {
     try {
         const id = req.params.id;
+
+        if(!id) return res.status(400).send({ message: "Id is required" });
       const user = await databaseInstance.db.User.findAll({
         where: {
           id: id,
@@ -28,6 +56,8 @@ exports.getUserById = async (req, res) => {
       res.status(200).send(user);
     } catch (error) {
       console.log(error);
+    handleApiError(res, error, "getUserById");
+
     }
   };
   
@@ -40,9 +70,8 @@ exports.getUserById = async (req, res) => {
       res.status(200).json({ message: "User created successfully" });
     } catch (error) {
       console.error(error);
+    handleApiError(res, error, "createUser");
       
-      // Send an error response
-      res.status(500).json({ error: "Internal server error" });
     }
   };
 
@@ -59,9 +88,9 @@ exports.getUserById = async (req, res) => {
       res.status(200).json({ message: "User edited successfully" });
     } catch (error) {
       console.error(error);
+    handleApiError(res, error, "editUser");
       
-      // Send an error response
-      res.status(500).json({ error: "Internal server error" });
+
     }
   };
 
@@ -69,7 +98,7 @@ exports.getUserById = async (req, res) => {
     try {
 
         const id = req.params.id;
-
+      if(!id) return res.status(400).send({ message: "Id is required" });
       const user = await databaseInstance.db.User.destroy( {
         where: {
           id: id,
@@ -81,9 +110,8 @@ exports.getUserById = async (req, res) => {
       res.status(200).json({ message: "User deleted successfully" });
     } catch (error) {
       console.error(error);
+    handleApiError(res, error, "deleteUser");
       
-      // Send an error response
-      res.status(500).json({ error: "Internal server error" });
     }
   };
